requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]



#-e .